refactor: add unit-createcomment module
