+++ core-commonui/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/di/Utils.kt
@@ -1,7 +1,6 @@
 package com.github.diegoberaldin.raccoonforlemmy.core.commonui.di
 
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.communitydetail.CommunityDetailMviModel
-import com.github.diegoberaldin.raccoonforlemmy.core.commonui.createcomment.CreateCommentMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.createpost.CreatePostMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.postdetail.PostDetailMviModel
 import com.github.diegoberaldin.raccoonforlemmy.core.commonui.userdetail.UserDetailMviModel
@@ -25,12 +24,6 @@ expect fun getUserDetailViewModel(
     user: UserModel,
     otherInstance: String = "",
 ): UserDetailMviModel
-
-expect fun getCreateCommentViewModel(
-    postId: Int? = null,
-    parentId: Int? = null,
-    editedCommentId: Int? = null,
-): CreateCommentMviModel
 
 expect fun getCreatePostViewModel(
     editedPostId: Int?,
