+++ core-commonui/unit-createcomment/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/createcomment/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.createcomment.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.createcomment.CreateCommentMviModel
+
+expect fun getCreateCommentViewModel(
+    postId: Int? = null,
+    parentId: Int? = null,
+    editedCommentId: Int? = null,
+): CreateCommentMviModel
