+++ core-commonui/unit-saveditems/build.gradle.kts
@@ -43,6 +43,7 @@ kotlin {
                 implementation(projects.coreCommonui.unitZoomableimage)
                 implementation(projects.coreCommonui.unitWeb)
                 implementation(projects.coreCommonui.unitCreatereport)
+                implementation(projects.coreCommonui.unitCreatecomment)
                 implementation(projects.coreCommonui.detailopenerApi)
                 implementation(projects.coreNavigation)
                 implementation(projects.corePersistence)
