+++ settings.gradle.kts
@@ -54,5 +54,6 @@ include(":core-commonui:unit-remove")
 include(":core-commonui:unit-reportlist")
 include(":core-commonui:unit-saveditems")
 include(":core-commonui:unit-createreport")
+include(":core-commonui:unit-createcomment")
 include(":core-commonui:detailopener-api")
 include(":core-commonui:detailopener-impl")
